+++ CHANGE_LOG.md
@@ -2,6 +2,16 @@
 
 - Alpha releases (working version, but API still evolving)
 
+## 0.6.0
+
+### Rule & Group Validation Configuration
+
+- Public interfaces have had the initial 'I' removed from their names.
+- An isPending property has been added to the root and group Validators which is true while validation is in progress.
+- Rule & Group configuration has been simplified.
+  - Intellisense for both is now fully working, both for configuration and in the output.
+  - Group configuration is now separate from the property rule definitions.
+
 ## 0.5.2
 
 ### Message Function